+++ colour/quality/__init__.py
@@ -106,7 +106,7 @@ def colour_fidelity_index(
     >>> from colour.colorimetry import SDS_ILLUMINANTS
     >>> sd = SDS_ILLUMINANTS["FL2"]
     >>> colour_fidelity_index(sd)  # doctest: +ELLIPSIS
-    70.1208244...
+    np.float64(70.1208244...)
     """
 
     method = validate_method(method, tuple(COLOUR_FIDELITY_INDEX_METHODS))
